remove FP sillyness from FindUnits & friends

+++ OpenRA.Game/Graphics/WorldRenderer.cs
@@ -58,8 +58,8 @@ namespace OpenRA.Graphics
 			bounds.Offset((int)Game.viewport.Location.X, (int)Game.viewport.Location.Y);
 
 			var actors = world.FindUnits(
-				new float2(bounds.Left, bounds.Top),
-				new float2(bounds.Right, bounds.Bottom));
+				new int2(bounds.Left, bounds.Top),
+				new int2(bounds.Right, bounds.Bottom));
 
 			var renderables = actors.SelectMany(a => a.Render())
 				.OrderBy(r => r, comparer);
