+++ OpenRA.Mods.RA/Combat.cs
@@ -80,7 +80,7 @@ namespace OpenRA.Mods.RA
 				case DamageModel.Normal:
 					{
 						var maxSpread = warhead.Spread * (float)Math.Log(Math.Abs(warhead.Damage), 2);
-						var hitActors = world.FindUnitsInCircle(args.dest, maxSpread);
+						var hitActors = world.FindUnitsInCircle(args.dest, (int)maxSpread);
 
 						foreach (var victim in hitActors)
 						{
@@ -92,7 +92,7 @@ namespace OpenRA.Mods.RA
 				case DamageModel.PerCell:
 					{
 						foreach (var t in world.FindTilesInCircle(targetTile, warhead.Size[0]))
-							foreach (var unit in world.FindUnits(Game.CellSize * t, Game.CellSize * (t + new float2(1,1))))
+							foreach (var unit in world.FindUnits(Game.CellSize * t, Game.CellSize * (t + new int2(1,1))))
 								unit.InflictDamage(args.firedBy,
 									(int)(warhead.Damage * warhead.EffectivenessAgainst(unit)), warhead);
 					} break;
